+++ src/components/ui/aside/index.ts
@@ -1,10 +1,5 @@
-import type { ComponentOptions } from 'vue';
-import Aside from './Aside.vue';
-
-export default {
-  install(app: ComponentOptions, { prefix }: { prefix: string }) {
-    app.component(`${prefix}Aside`, Aside);
-  },
-};
-
-export { Aside };
+export { default as Aside } from './Aside.vue';
+export { default as AsideBody } from './AsideBody.vue';
+export { default as AsideFooter } from './AsideFooter.vue';
+export { default as AsideSection } from './AsideSection.vue';
+export type { AsideProps, AsideSectionProps } from './types';
