+++ src/index.ts
@@ -41,6 +41,7 @@ export * from './components/admin/user-profile-card';
 
 // UI
 export * from './components/ui/alert';
+export * from './components/ui/aside';
 export * from './components/ui/badge';
 export * from './components/ui/dropdown';
 export * from './components/ui/form-select';
@@ -57,8 +58,6 @@ export { default as Stack } from './components/ui/stack/Stack.vue';
 export { default as Button } from './components/ui/button/Button.vue';
 export { default as ButtonGroup } from './components/ui/button-group/ButtonGroup.vue';
 
-export { default as Aside } from './components/ui/aside/Aside.vue';
-export { default as AsideSection } from './components/ui/aside/AsideSection.vue';
 export { default as Badge } from './components/ui/badge/Badge.vue';
 export { default as FormCheckbox } from './components/ui/form-checkbox/FormCheckbox.vue';
 export { default as FormTextfield } from './components/ui/form-textfield/FormTextfield.vue';
